Updated eslint deps, cleaned up inputdisplay, default export
