+++ .eslintrc.json
@@ -1,3 +1,16 @@
 {
-    "extends": "airbnb"
+  "parser": "babel-eslint",
+  "extends": "airbnb",
+  "env": {
+    "browser": true,
+    "node": true,
+    "es6": true
+  },
+  "ecmaFeatures": {
+    "modules": true,
+    "jsx": true
+  },
+  "rules": {
+    "no-undef": 0
+  }
 }
