+++ package.json
@@ -5,6 +5,9 @@
   "homePage": "https://sstpierre2.github.io/Nice-Calculator",
   "dependencies": {
     "bootstrap": "4.0.0",
+    "jquery": "1.9.1",
+    "popper.js": "^1.12.9",
+    "prop-types": "^15.6.1",
     "react": "^16.3.1",
     "react-dom": "^16.3.1",
     "react-scripts": "1.1.4"
@@ -18,10 +21,11 @@
   "devDependencies": {
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.10.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0",
+    "eslint": "3.19.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^5.0.1",
+    "eslint-plugin-react": "^7.1.0",
     "firebase-tools": "^3.18.4",
     "react-test-renderer": "^16.3.1"
   }
